Improved to 64-bit numerical precision

+++ projects/mandelbrot/src/julia_set/mod.rs
@@ -7,18 +7,18 @@ mod presets;
 
 #[derive(Copy, Clone, Debug)]
 pub struct JuliaSet {
-    pub center: Complex<f32>,
+    pub center: Complex<f64>,
     pub max_iterations: u32,
-    pub tint_magnification: f32,
-    pub shift: Complex<f32>,
-    pub power: f32,
-    pub zoom: f32,
+    pub tint_magnification: f64,
+    pub shift: Complex<f64>,
+    pub power: f64,
+    pub zoom: f64,
 }
 
 
 
 impl JuliaSet {
-    pub fn escape(&self, mut z: Complex<f32>) -> EscapeSpeed {
+    pub fn escape(&self, mut z: Complex<f64>) -> EscapeSpeed {
         let mut i = 0;
         while i < self.max_iterations && z.norm() < 32.0 {
             z = z.powf(self.power) + self.shift;
