+++ projects/mandelbrot/src/mandelbrot_set/mod.rs
@@ -6,11 +6,11 @@ use crate::palette::EscapeSpeed;
 
 #[derive(Copy, Clone, Debug)]
 pub struct MandelbrotSet {
-    pub center: Complex<f32>,
+    pub center: Complex<f64>,
     pub max_iterations: u32,
-    pub tint_magnification: f32,
-    pub power: f32,
-    pub zoom: f32,
+    pub tint_magnification: f64,
+    pub power: f64,
+    pub zoom: f64,
 }
 
 impl Default for MandelbrotSet {
@@ -20,41 +20,41 @@ impl Default for MandelbrotSet {
 }
 
 impl MandelbrotSet {
-    pub fn with_center(self, x: f32, y: f32) -> Self {
+    pub fn with_center(self, x: f64, y: f64) -> Self {
         Self { center: Complex::new(x, y), ..self }
     }
     pub fn with_max_iterations(self, iterations: u32) -> MandelbrotSet {
         Self { max_iterations: iterations, ..self }
     }
-    pub fn with_tint_coefficient(self, tint_coefficient: f32) -> MandelbrotSet {
+    pub fn with_tint_coefficient(self, tint_coefficient: f64) -> MandelbrotSet {
         Self { tint_magnification: tint_coefficient, ..self }
     }
-    pub fn with_zoom(self, zoom: f32) -> MandelbrotSet {
+    pub fn with_zoom(self, zoom: f64) -> MandelbrotSet {
         Self { zoom, ..self }
     }
 }
 
 
 impl CanvasRenderer for MandelbrotSet {
-    fn center_x(&self) -> f32 {
+    fn center_x(&self) -> f64 {
         self.center.re
     }
 
-    fn center_y(&self) -> f32 {
+    fn center_y(&self) -> f64 {
         self.center.im
     }
 
-    fn zoom_reciprocal(&self) -> f32 {
+    fn zoom_reciprocal(&self) -> f64 {
         self.zoom.recip()
     }
 
-    fn render_pixel(&self, c: Complex<f32>) -> Rgba<u8> {
-        self.escape(c).tint_by_max((self.max_iterations as f32).div(self.tint_magnification))
+    fn render_pixel(&self, c: Complex<f64>) -> Rgba<u8> {
+        self.escape(c).tint_by_max((self.max_iterations as f64).div(self.tint_magnification))
     }
 }
 
 impl MandelbrotSet {
-    pub fn escape(&self, c: Complex<f32>) -> EscapeSpeed {
+    pub fn escape(&self, c: Complex<f64>) -> EscapeSpeed {
         let mut z = Complex::default();
         let mut i = 0;
         while i < self.max_iterations && z.norm() < 32.0 {
